add: 施設予約, タイトル指定の対応
